// src/Scrapers/Mizrahi/Interfaces/ScrapedTransaction.ts
export interface IScrapedTransaction {
  RecTypeSpecified: boolean;
  MC02PeulaTaaEZ: string;
  MC02SchumEZ: number;
  MC02AsmahtaMekoritEZ: string;
  MC02TnuaTeurEZ: string;
  IsTodayTransaction: boolean;
  MC02ErehTaaEZ: string;
  MC02ShowDetailsEZ?: string;
  MC02KodGoremEZ: string;
  MC02SugTnuaKaspitEZ: string;
  MC02AgidEZ: string;
  MC02SeifMaralEZ: string;
  MC02NoseMaralEZ: string;
  TransactionNumber: string | number;
}